+++ app/javascript/packs/application.js
@@ -8,33 +8,6 @@ require("@rails/ujs").start()
 require("turbolinks").start()
 require("@rails/activestorage").start()
 require("channels")
-
-import mapboxgl from 'mapbox-gl';
-const mapElement = document.getElementById('map');
-
-mapboxgl.accessToken = mapElement.dataset.mapboxApiKey;
-const map = new mapboxgl.Map({
-  container: 'map',
-  style: 'mapbox://styles/mapbox/streets-v11'
-});
-
-
-const markers = JSON.parse(mapElement.dataset.markers);
-
-markers.forEach((marker) => {
-  new mapboxgl.Marker()
-    .setLngLat([ marker.lng, marker.lat ])
-    .addTo(map);
-});
-
-
-
-
-
-
-
-
-
 // Uncomment to copy all static images under ../images to the output folder and reference
 // them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
 // or the `imagePath` JavaScript helper below.
@@ -50,13 +23,14 @@ markers.forEach((marker) => {
 
 // External imports
 import "bootstrap";
-
+import { initMapbox } from '../plugins/init_mapbox';
 // Internal imports, e.g:
 // import { initSelect2 } from '../components/init_select2';
 
 document.addEventListener('turbolinks:load', () => {
   // Call your functions here, e.g:
   // initSelect2();
+  initMapbox();
 });
